Implement a config class for better configuration
The idea is to create a configuration from a config file (or manually using a sequential API) and then start the server using that configuration. Further down the line the idea is also to enable the user to preprocess all the data before even starting the server and passing only the preprocessed data into the server instance for faster server development. The tight data processing + server initialization can be painfully slow at times.